fix(de): View Type Dropdown will use the defined types (#191)
* chore(visualization): move all the visualization types from flows to a shared folder

* feat(de): update dataexplorer view type

* chore(visualization): move function for visualization to shared folder

* fix(visualization): move view type dropdown component to shared folder
