+++ src/dashboards/components/VEOHeader.tsx
@@ -11,7 +11,7 @@ import {
   Page,
 } from '@influxdata/clockface'
 import VisOptionsButton from 'src/timeMachine/components/VisOptionsButton'
-import ViewTypeDropdown from 'src/timeMachine/components/view_options/ViewTypeDropdown'
+import Visualization from 'src/timeMachine/components/visualization/view'
 
 // Constants
 import {
@@ -44,7 +44,7 @@ class VEOHeader extends PureComponent<Props> {
         </Page.Header>
         <Page.ControlBar fullWidth={true}>
           <Page.ControlBarLeft>
-            <ViewTypeDropdown />
+            <Visualization />
             <VisOptionsButton />
           </Page.ControlBarLeft>
           <Page.ControlBarRight>
