+++ src/shared/visualization/types/Band/icon.tsx
