+++ src/shared/visualization/types/Gauge/icon.tsx
