+++ src/shared/visualization/types/Graph/properties.ts
