+++ src/shared/visualization/types/Heatmap/properties.ts
