+++ src/shared/visualization/types/Histogram/icon.tsx
