+++ src/shared/visualization/types/Histogram/properties.ts
