+++ src/shared/visualization/types/Map/icon.tsx
