+++ src/shared/visualization/types/Mosaic/index.tsx
