+++ src/shared/visualization/types/Scatter/icon.tsx
