+++ src/shared/visualization/types/Scatter/index.tsx
