+++ src/shared/visualization/types/Scatter/properties.ts
