+++ src/shared/visualization/types/SingleStat/SingleStatOptions.tsx
