+++ src/shared/visualization/types/SingleStat/properties.ts
