+++ src/shared/visualization/types/SingleStatWithLine/properties.ts
