+++ src/shared/visualization/types/Table/TableOptions.tsx
