+++ src/shared/visualization/types/Table/index.tsx
